move types and extract functions
