+++ src/Frontend/src/components/messages/MessageView.vue
@@ -14,7 +14,8 @@ import EditRetryDialog from "../failedmessages/EditRetryDialog.vue";
 import routeLinks from "@/router/routeLinks";
 import { EditAndRetryConfig } from "@/resources/Configuration";
 import { TYPE } from "vue-toastification";
-import { ExtendedFailedMessage, FailedMessageError, FailedMessageStatus, isError } from "@/resources/FailedMessage";
+import { FailedMessageError, FailedMessageStatus } from "@/resources/FailedMessage";
+import { ExtendedFailedMessage, isError } from "../messages2/FailedMessage";
 import Message from "@/resources/Message";
 import { NServiceBusHeaders } from "@/resources/Header";
 import { useConfiguration } from "@/composables/configuration";
