+++ src/Frontend/src/resources/FailedMessage.ts
@@ -1,6 +1,4 @@
 import type EndpointDetails from "@/resources/EndpointDetails";
-import type Header from "@/resources/Header";
-import { ConversationModel } from "./SequenceDiagram/SequenceModel";
 
 export default interface FailedMessage {
   id: string;
@@ -20,39 +18,9 @@ export default interface FailedMessage {
   edit_of: string;
 }
 
-export interface ExtendedFailedMessage extends FailedMessage {
-  error_retention_period: number;
-  delete_soon: boolean;
-  deleted_in: string;
-  retryInProgress: boolean;
-  deleteInProgress: boolean;
-  restoreInProgress: boolean;
-  selected: boolean;
-  retried: boolean;
-  archiving: boolean;
-  restoring: boolean;
-  archived: boolean;
-  resolved: boolean;
-  headersNotFound: boolean;
-  messageBodyNotFound: boolean;
-  bodyUnavailable: boolean;
-  headers: Header[];
-  conversationId: string;
-  conversation?: ConversationModel;
-  messageBody: string;
-  contentType: string;
-  isEditAndRetryEnabled: boolean;
-  redirect: boolean;
-  submittedForRetrial: boolean;
-}
-
 export interface FailedMessageError {
   notFound: boolean;
   error: boolean;
-}
-
-export function isError(obj: ExtendedFailedMessage | FailedMessageError): obj is FailedMessageError {
-  return (obj as FailedMessageError).error !== undefined || (obj as FailedMessageError).notFound !== undefined;
 }
 
 export interface ExceptionDetails {
